src/main/webapp/app/entities/slider/delete/slider-delete-dialog.component.ts
import {Component} from '@angular/core';
import {FormsModule} from '@angular/forms';
import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';

import SharedModule from 'app/shared/shared.module';
import {ISlider} from '../slider.model';
import {SliderService} from '../service/slider.service';
import {ITEM_DELETED_EVENT} from 'app/config/navigation.constants';

@Component({
  standalone: true,
  templateUrl: './slider-delete-dialog.component.html',
  imports: [SharedModule, FormsModule],
})
export class SliderDeleteDialogComponent {
  slider?: ISlider;

  constructor(protected sliderService: SliderService, protected activeModal: NgbActiveModal) {
  }

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: number): void {
    this.sliderService.delete(id).subscribe(() => {
      this.activeModal.close(ITEM_DELETED_EVENT);
    });
  }
}
